fix: `SkelModule.default_order` generalized
- fixes that `getList()` in Jinja-templates doesn't apply the default_order
- removes rendundant codebases
- I don't know why I didn't do it this way before?
