+++ src/viur/core/prototypes/list.py
@@ -5,7 +5,7 @@ from viur.core.decorators import *
 from viur.core.cache import flushCache
 from viur.core.skeleton import SkeletonInstance
 from viur.core.bones import BaseBone
-from .skelmodule import SkelModule, DEFAULT_ORDER_TYPE
+from .skelmodule import SkelModule
 
 
 class List(SkelModule):
@@ -19,14 +19,6 @@ class List(SkelModule):
     """
     handler = "list"
     accessRights = ("add", "edit", "view", "delete", "manage")
-
-    default_order: DEFAULT_ORDER_TYPE = None
-    """
-    Allows to specify a default order for this module, which is applied when no other order is specified.
-
-    Setting a default_order might result in the requirement of additional indexes, which are being raised
-    and must be specified.
-    """
 
     def viewSkel(self, *args, **kwargs) -> SkeletonInstance:
         """
@@ -207,39 +199,7 @@ class List(SkelModule):
         if not (query := self.listFilter(self.viewSkel().all().mergeExternalFilter(kwargs))):
             raise errors.Unauthorized()
 
-        # Apply default_order when possible!
-        # TODO: refactor: Duplicate code in prototypes.Tree
-        if (
-                self.default_order
-                and query.queries
-                and not isinstance(query.queries, list)
-                and not query.queries.orders
-                and not current.request.get().kwargs.get("search")
-        ):
-            if callable(default_order := self.default_order):
-                default_order = default_order(query)
-
-            if isinstance(default_order, dict):
-                logging.debug(f"Applying filter {default_order=}")
-                query.mergeExternalFilter(default_order)
-
-            elif default_order:
-                logging.debug(f"Applying {default_order=}")
-
-                # FIXME: This ugly test can be removed when there is type that abstracts SortOrders
-                if (
-                    isinstance(default_order, str)
-                    or (
-                        isinstance(default_order, tuple)
-                        and len(default_order) == 2
-                        and isinstance(default_order[0], str)
-                        and isinstance(default_order[1], db.SortOrder)
-                    )
-                ):
-                    query.order(default_order)
-                else:
-                    query.order(*default_order)
-
+        self._apply_default_order(query)
         return self.render.list(query.fetch())
 
     @force_ssl
